Extract helpers to their own file
