+++ lib/flutter_layout_grid.dart
@@ -1,93 +1,7 @@
 library flutter_layout_grid;
 
-import 'package:flutter/widgets.dart';
-import 'package:flutter_layout_grid/src/widgets/placement.dart';
-import 'package:meta/meta.dart';
-
-import 'src/rendering/track_size.dart';
-
 export 'src/foundation/placement.dart';
+export 'src/helpers.dart';
 export 'src/rendering/track_size.dart';
 export 'src/widgets/layout_grid.dart';
 export 'src/widgets/placement.dart';
-export 'src/widgets/placement.dart';
-
-// Helpers
-
-/// An [IntrinsicContentTrackSize], mirroring CSS's name for the track sizing
-/// function.
-const auto = IntrinsicContentTrackSize();
-
-/// Returns a track size that is sized based on its contents.
-IntrinsicContentTrackSize intrinsic({String debugLabel}) =>
-    IntrinsicContentTrackSize(debugLabel: debugLabel);
-
-/// Returns a new track size that is exactly [sizeInPx] wide.
-FixedTrackSize fixed(double sizeInPx, {String debugLabel}) =>
-    FixedTrackSize(sizeInPx, debugLabel: debugLabel);
-
-/// Returns a new track size that expands to fill available space.
-FlexibleTrackSize flex(double flexFactor, {String debugLabel}) =>
-    FlexibleTrackSize(flexFactor, debugLabel: debugLabel);
-
-/// Defines a set of extension methods on [num] for creating tracks
-extension TrackUnitsNumExtension on num {
-  FixedTrackSize get px => fixed(toDouble());
-  FlexibleTrackSize get fr => flex(toDouble());
-}
-
-/// Returns this list repeated [times] times.
-///
-///     repeat(2, [fixed(100), fixed(200)])
-///     // [fixed(100), fixed(200), fixed(100), fixed(200)]
-///
-List<TrackSize> repeat(int times, List<TrackSize> tracks) =>
-    _repeat(times, tracks).toList();
-
-Iterable<T> _repeat<T>(int times, Iterable<T> source) sync* {
-  for (int i = 0; i < times; i++) {
-    yield* source;
-  }
-}
-
-/// Convenience function for pretty grid definitions, ala:
-///
-///     LayoutGrid(
-///       templateAreas: gridTemplateAreas([
-///         'pink   pink   .',
-///         'pink   pink   red',
-///         'orange yellow red',
-///       ]),
-///       templateColumnSizes: [fixed(100), fixed(100), fixed(100)],
-///       templateRowSizes: [
-///         fixed(100),
-///         fixed(100),
-///         fixed(100),
-///        ],
-///       children: [
-///         //!!!
-///         gridArea('pink').containing(Container(color: Colors.pink)),
-///         gridArea('red').containing(Container(color: Colors.red)),
-///         gridArea('orange').containing(Container(color: Colors.orange)),
-///         gridArea('yellow').containing(Container(color: Colors.yellow)),
-///       ],
-///     )
-///
-NamedAreaGridPlacementBuilder gridArea(String name) {
-  return NamedAreaGridPlacementBuilder._(name);
-}
-
-/// Dumb little helper class for slightly cleaner grid child placement
-@immutable
-class NamedAreaGridPlacementBuilder {
-  NamedAreaGridPlacementBuilder._(this.name);
-
-  final String name;
-
-  NamedAreaGridPlacement containing(Widget child) {
-    return NamedAreaGridPlacement(
-      areaName: name,
-      child: child,
-    );
-  }
-}
